refactor(python): split runner flush tests by protocol lifecycle
